+++ jest.integration.config.js
@@ -1,0 +1,6 @@
+import baseConfig from './jest.config.js'
+export default {
+  ...baseConfig,
+  testMatch: ['<rootDir>/test/integration/**/*.test.[jt]s?(x)'],
+  coverageThreshold: {},
+}