+++ test/integration/sync/repositories/progress.test.ts
@@ -4,7 +4,7 @@ jest.mock('@/services/sync/manager', () => ({ default: class SyncManager {} }))
 jest.mock('@/services/sync/repository-proxy', () => ({ db: {} }))
 
 import { Database } from '@nozbe/watermelondb'
-import { makeDatabase, makeStore, resetDatabase } from '../helpers/index'
+import { makeDatabase, makeStore, resetDatabase } from '../../../unit/sync/helpers/index'
 import ContentProgress, {
   COLLECTION_TYPE,
   COLLECTION_ID_SELF,